requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0
openai>=1.12.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
requests==2.31.0
